# lib/api-zod/src/generated/types/productSearchResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Company } from "./company";
import type { Country } from "./country";
import type { GlobalStats } from "./globalStats";

export interface ProductSearchResult {
  productName: string;
  producers: Company[];
  distributors: Company[];
  exportingCountries: Country[];
  importingCountries: Country[];
  globalStats: GlobalStats;
  /** Approximate year of data */
  dataYear: string;
}